+++ clumio/plugin_framework/clumio_gcp_connection/consts.go
@@ -1,0 +1,14 @@
+// Copyright (c) 2025 Clumio, a Commvault Company All Rights Reserved
+
+package clumio_gcp_connection
+
+const (
+	// Constants used by the resource model for the clumio_gcp_connection Terraform resource. These
+	// values should match the schema tfsdk tags on the resource model struct in schema.go.
+	schemaID                     = "id"
+	schemaClumioControlPlaneId   = "clumio_control_plane_id"
+	schemaClumioControlPlaneRole = "clumio_control_plane_role"
+	schemaProjectId              = "project_id"
+	schemaDescription            = "description"
+	schemaToken                  = "token"
+)